+++ packages/tailwindcss-language-service/src/util/v4/design-system.ts
@@ -2,7 +2,11 @@ import postcss from 'postcss'
 import type { Rule } from './ast'
 import type { NamedVariant } from './candidate'
 
-export interface Theme {}
+export interface Theme {
+  // Prefix didn't exist on
+  prefix?: string
+  entries(): [string, any][]
+}
 
 export interface ClassMetadata {
   modifiers: string[]